Added a check in Robot.cpp that makes sure it only runs the code for moving the spatula only once per button press
This is done by having the spatulaMoveButtonPressed boolean keep track
of whether button 7 on the joystick is pressed and the end of each loop.
If the button wasn't pressed the last loop, but is pressed in the
current loop, the code will execute.

Also added code that will toggle the spatulaUp boolean.